changed the amount datatype to double everywhere i found and updated the enum files

+++ src/main/java/com/asu/cse545/group12/services/AccountService.java
@@ -9,13 +9,13 @@ import com.asu.cse545.group12.domain.Users;
 public interface AccountService {
 
 	public int insertRow(int userId);
-	public boolean isBalanceValid(int accountNumber, int amount,String type);
+	public boolean isBalanceValid(int accountNumber, double amount,String type);
 	public boolean isValidAccountNumber(int accountNumber);
 	public boolean isAccountNumberBelongToUser(int accountNumber, String userName);
 	public Account getAccount(String username);
 	public Account getAccount(int accountNumber);
-	public boolean doCredit(int accountNumber, int amount);
-	public boolean doDebit(int accountNumber, int amount);
+	public boolean doCredit(int accountNumber, double amount);
+	public boolean doDebit(int accountNumber, double amount);
 	public Account getRowById(int accountNumber);
 	public int updateRow(Account account);
 	public List<Account> getAccounts(int userId);
